Title: Cleans and Normalizes DBF files obtained from experiments using the FLUOstar microplate reader.

Input the path to a ".dbf" file obtained from the FLUOstar microplate (usually a 96-well microplate) reader; this function will create a data frame, clean the data frame, normalize the data frame, append a "Cycle_Number" column and return a data frame that is ready for analysis. Most importantly, this function is a single_step function. Also, the function can be extended to other ".dbf" files if they follow the format for which this function was designed; this is totally at the users' discretion.

Usage

normfluordbf(file = NULL, norm_scale = NULL, transformed = NULL, fun = NA, ...)

Arguments

file

A string ("liposomes_xxx.dbf") if the file is found within the present working directory (pwd) OR a path pointing directly to a ".dbf" file, from FLUOstar experiments.

norm_scale

This parameter can taken in 'hundred', 'one', or 'z-score' which denotes the normalization type; Initialized as NULL.

transformed

This parameter can take in 'log' which denotes a logarithmic box-cox transformation; Initialized as NULL.

fun

A variable defined as NA, used for boolean expressions or manipulation.

...

A container object that can be used to capture extra variables if needed.

Value

A normalized data frame with an appended "Cycle_Number" attribute.

Note

Re-nomenclature of norm_tidy_dbf to a more appropriate name that facilitates function utilization. Users can continue with the old name ("norm_tidy_dbf") but this is a better name in my opinion.

Author

Tingwei Adeck

Examples

fpath <- system.file("extdata", "liposomes_214.dbf", package = "normfluodbf", mustWork = TRUE)
normalized_dbf <- normfluordbf(file=fpath, norm_scale = 'raw')
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 1
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 1
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 1
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 1
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 1
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 1
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 1
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 1
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 1
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 2
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 2
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 2
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 2
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 2
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 2
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 2
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 2
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 2
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 3
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 3
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 3
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 3
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 3
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 3
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 3
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 3
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 3
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 4
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 4
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 4
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 4
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 4
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 4
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 4
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 4
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 4
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 5
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 5
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 5
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 5
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 5
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 5
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 5
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 5
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 5
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 6
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 6
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 6
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 6
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 6
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 6
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 6
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 6
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 6
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 7
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 7
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 7
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 7
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 7
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 7
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 7
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 7
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 7
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 8
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 8
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 8
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 8
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 8
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 8
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 8
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 8
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 8
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 9
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 9
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 9
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 9
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 9
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 9
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 9
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 9
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 9
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 10
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 10
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 10
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 10
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 10
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 10
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 10
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 10
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 10
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 11
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 11
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 11
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 11
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 11
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 11
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 11
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 11
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 11
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 12
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 12
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 12
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 12
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 12
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 12
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 12
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 12
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 12
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 13
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 13
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 13
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 13
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 13
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 13
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 13
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 13
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 13
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 14
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 14
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 14
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 14
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 14
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 14
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 14
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 14
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 14
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 15
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 15
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 15
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 15
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 15
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 15
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 15
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 15
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 15
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 16
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 16
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 16
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 16
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 16
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 16
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 16
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 16
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 16
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 17
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 17
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 17
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 17
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 17
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 17
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 17
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 17
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 17
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 18
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 18
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 18
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 18
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 18
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 18
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 18
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 18
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 18
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 19
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 19
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 19
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 19
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 19
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 19
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 19
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 19
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 19
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 20
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 20
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 20
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 20
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 20
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 20
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 20
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 20
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 20
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 21
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 21
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 21
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 21
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 21
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 21
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 21
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 21
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 21
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 22
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 22
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 22
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 22
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 22
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 22
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 22
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 22
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 22
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 23
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 23
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 23
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 23
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 23
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 23
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 23
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 23
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 23
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 24
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 24
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 24
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 24
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 24
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 24
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 24
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 24
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 24
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 25
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 25
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 25
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 25
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 25
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 25
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 25
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 25
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 25
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 26
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 26
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 26
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 26
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 26
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 26
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 26
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 26
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 26
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 27
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 27
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 27
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 27
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 27
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 27
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 27
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 27
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 27
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 28
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 28
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 28
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 28
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 28
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 28
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 28
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 28
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 28
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 29
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 29
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 29
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 29
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 29
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 29
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 29
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 29
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 29
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 30
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 30
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 30
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 30
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 30
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 30
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 30
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 30
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 30
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 31
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 31
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 31
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 31
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 31
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 31
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 31
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 31
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 31
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 32
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 32
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 32
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 32
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 32
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 32
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 32
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 32
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 32
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 33
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 33
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 33
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 33
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 33
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 33
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 33
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 33
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 33
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 34
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 34
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 34
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 34
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 34
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 34
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 34
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 34
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 34
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 35
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 35
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 35
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 35
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 35
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 35
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 35
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 35
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 35
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 36
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 36
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 36
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 36
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 36
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 36
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 36
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 36
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 36
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 37
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 37
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 37
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 37
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 37
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 37
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 37
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 37
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 37
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 38
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 38
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 38
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 38
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 38
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 38
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 38
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 38
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 38
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 39
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 39
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 39
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 39
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 39
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 39
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 39
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 39
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 39
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 1 row: 40
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 2 row: 40
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 3 row: 40
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 4 row: 40
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 5 row: 40
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 6 row: 40
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 7 row: 40
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 8 row: 40
#> ğŸ·ğŸ«ğŸ¦
#> YIKES, value > 2^15, Watch in future experimental designs column: 9 row: 40